[pipes, tests] add common utils.py to Pipes tests
