+++ python_modules/dagster/dagster_tests/execution_tests/pipes_tests/utils.py
@@ -1,0 +1,15 @@
+import inspect
+import textwrap
+from contextlib import contextmanager
+from tempfile import NamedTemporaryFile
+from typing import Any, Callable, Iterator
+
+
+@contextmanager
+def temp_script(script_fn: Callable[[], Any]) -> Iterator[str]:
+    # drop the signature line
+    source = textwrap.dedent(inspect.getsource(script_fn).split("\n", 1)[1])
+    with NamedTemporaryFile() as file:
+        file.write(source.encode())
+        file.flush()
+        yield file.name
